feat: creating slider component
