+++ src/pages/product/productSlider/slider.scss
